+++ constcheck/__main__.py
@@ -4,7 +4,9 @@ constcheck.__main__
 
 Module entry point.
 """
+import sys as _sys
+
 from constcheck import main
 
 if __name__ == "__main__":
-    main()
+    _sys.exit(main())